+++ src/test/attrs.test.js
@@ -198,4 +198,32 @@ describe('attrs', () => {
     })``;
     expect(TestRenderer.create(<StyledComp />).toJSON()).toMatchSnapshot();
   });
+
+  describe('warnings', () => {
+    beforeEach(() => {
+      jest.spyOn(console, 'warn').mockImplementation(() => {});
+    });
+
+    it('warns upon use of a Stateless Functional Component as a prop', () => {
+      const Inner = () => <div />;
+      const Comp = styled.div.attrs({ component: Inner })``;
+
+      TestRenderer.create(<Comp />);
+
+      expect(console.warn.mock.calls[0][0]).toMatchInlineSnapshot(`
+"It looks like you've used a component as value for the component prop in the attrs constructor.
+You'll need to wrap it in a function to make it available inside the styled component.
+For example, { component: () => InnerComponent } instead of { component: InnerComponent }"
+`);
+    });
+
+    it('does not warn if the Stateless Functional Component is wrapped in a function', () => {
+      const Inner = () => <div />;
+      const Comp = styled.div.attrs({ component: () => Inner })``;
+
+      TestRenderer.create(<Comp />);
+
+      expect(console.warn).not.toHaveBeenCalled();
+    });
+  });
 });